backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
faster-whisper>=1.0.0
piper-tts>=1.2.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
